requirements.txt
Flask
line-bot-sdk==3.7.0
google-cloud-storage>=2.17.0
gunicorn>=21.2.0
